requirements.txt
Django==1.7.4
Sphinx==1.2.3
coverage==3.7.1
django-haystack==2.3.1
djangorestframework==3.0.5
elasticsearch==1.4.0
fake-factory==0.4.2
mock==1.0.1
nose==1.3.4
sphinx-rtd-theme==0.1.6
urllib3==1.10.1
